requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Spark and big data processing
pyspark==3.5.0
py4j==0.10.9.7

# PostgreSQL connectivity
psycopg2-binary==2.9.9
apache-airflow-providers-postgres==5.8.0

# Additional utilities for data processing - compatible versions
pandas==2.1.4
numpy>=1.26.0,<2.0.0

# Logging and monitoring - compatible with Airflow 3.0.3
structlog>=25.4.0

apache-airflow-providers-postgres>=5.8.0